src/containers/docker/graphql/backend/requirements.txt
fastapi
graphene
graphene-sqlalchemy
pydantic
sqlalchemy
alembic
strawberry-graphql[fastapi]
uvicorn[standard]
asyncpg
sqlalchemy[asyncio]
asyncpg
websockets
aiokafka
aioredis
